sclab_sites/oncotyping/db_create.py
#!/usr/bin/env python

import sys
import os
basedir = os.path.dirname(os.path.abspath(os.path.dirname(__file__)))
sys.path.append(os.path.dirname(basedir))
print sys.path
from migrate.versioning import api
from config import SQLALCHEMY_DATABASE_URI
from config import SQLALCHEMY_MIGRATE_REPO
from sclab_sites.oncotyping.onco_views import oncotypingdb as db
from sclab_sites.oncotyping.onco_models import User

db.create_all()

if not os.path.exists(SQLALCHEMY_MIGRATE_REPO):
    api.create(SQLALCHEMY_MIGRATE_REPO, 'database repository')
    api.version_control(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO)
else:
    api.version_control(SQLALCHEMY_DATABASE_URI, SQLALCHEMY_MIGRATE_REPO, api.version(SQLALCHEMY_MIGRATE_REPO))

fake_login = User('1', 'test', 'test', 'test@example.com')
db.session.add(fake_login)
db.session.commit()