sclab_sites/requirements.txt
blinker==1.3
decorator==3.4.0
Flask==0.10.1
Flask-Admin==1.1.0
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Sijax==0.4.1
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.3
future==0.14.3
gunicorn==19.1.1
itsdangerous==0.24
Jinja2==2.7.3
MarkupSafe==0.23
pbr==0.10.3
PyMySQL==0.6.3
Sijax==0.3.2
six==1.8.0
SQLAlchemy==0.9.8
sqlalchemy-migrate==0.9.2
sqlparse==0.1.14
Tempita==0.5.2
Werkzeug==0.9.6
WTForms==2.0.1
